requirements.txt
ultralytics>=8.0.20
streamlit>=1.25.0
opencv-python
pillow
numpy
